story api config model and service

+++ Backend/src/5-logic/commemoration-logic-area/story-logic.ts
@@ -1,5 +1,6 @@
 import dal from "../../2-utils/dal";
 import StoryModel from "../../4-models/commemorations-models/story-model";
+import { ValidationErrorModel } from "../../4-models/error-models";
 
 async function getCommemorativeStories(commemorativeID: number) {
     const sql = `SELECT * FROM stories WHERE commemorativeID = ?`;
@@ -8,15 +9,18 @@ async function getCommemorativeStories(commemorativeID: number) {
 }
 
 async function addCommemorativeStory(story: StoryModel) {
-    const sql = `INSERT INTO stories VALUES (DEFAULT, ?, ?, ?)`;
-    const info = await dal.execute(sql, [story.commemorativeID, story.author, story.story]);
+    const err = story.validation();
+    if (err) throw new ValidationErrorModel(err);
+
+    const sql = `INSERT INTO stories VALUES (DEFAULT, ?, ?, ?, ?)`;
+    const info = await dal.execute(sql, [story.commemorativeID, story.userID, story.author, story.story]);
     story.storyID = info.insertId;
     return story;
 }
 
-async function deleteCommemorativeStory(storyID: number) {
-    const sql = `DELETE FROM stories WHERE storyID = ?`;
-    await dal.execute(sql, [storyID]);
+async function deleteCommemorativeStory(storyID: number, userID: number) {
+    const sql = `DELETE FROM stories WHERE storyID = ? AND userID = ?`;
+    await dal.execute(sql, [storyID, userID]);
 }
 
 export default { getCommemorativeStories, addCommemorativeStory, deleteCommemorativeStory };
